move trie inside main migration
