+++ migration/state/newstate/internal/trie/counter.go
